Edited Impersonation workflow

+++ routes/content-pages.php
@@ -9,12 +9,12 @@ Route::get('/pages/{contentPage:slug}', [ContentPageController::class, 'show'])
 
 Route::prefix('admin/content-pages')
     ->name('admin.content-pages.')
-    ->middleware('permission:view_admin')
+    ->middleware(['permission:view_admin', 'permission:access_content_pages'])
     ->group(function () {
         Route::get('/', [AdminContentPageController::class, 'index'])->name('index');
-        Route::get('/create', [AdminContentPageController::class, 'create'])->name('create');
-        Route::post('/', [AdminContentPageController::class, 'store'])->name('store');
-        Route::get('/{contentPage}/edit', [AdminContentPageController::class, 'edit'])->name('edit');
-        Route::put('/{contentPage}', [AdminContentPageController::class, 'update'])->name('update');
-        Route::delete('/{contentPage}', [AdminContentPageController::class, 'destroy'])->name('destroy');
+        Route::get('/create', [AdminContentPageController::class, 'create'])->middleware('permission:manage_content_pages')->name('create');
+        Route::post('/', [AdminContentPageController::class, 'store'])->middleware('permission:manage_content_pages')->name('store');
+        Route::get('/{contentPage}/edit', [AdminContentPageController::class, 'edit'])->middleware('permission:manage_content_pages')->name('edit');
+        Route::put('/{contentPage}', [AdminContentPageController::class, 'update'])->middleware('permission:manage_content_pages')->name('update');
+        Route::delete('/{contentPage}', [AdminContentPageController::class, 'destroy'])->middleware('permission:manage_content_pages')->name('destroy');
     });
